\docType{package}
\name{multistep}
\alias{multistep}
\alias{multistep-package}
\title{Utilities for running, timing and logging a multi-step process from R.}
\description{
\tabular{ll}{ Package: \tab multistep\cr Type: \tab
Package\cr Version: \tab 0.1\cr Date: \tab 2014-05-26\cr
License: \tab Apache License, Version 2.0\cr Author: \tab
Sergei Izrailev, Jeremy Stanley\cr Maintainer: \tab Sergei
Izrailev <sizrailev@collective.com>\cr LazyLoad: \tab
yes\cr }
}

